pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "content-factory"
version = "0.1.0"
description = "Reproducible content-writing pipeline for Blog, LinkedIn, and X/Twitter"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9,<1",
    "pydantic>=2,<3",
    "pyyaml>=6,<7",
    "rich>=13,<14",
    "langgraph>=0.2",
    "openai>=1,<2",
]

[project.optional-dependencies]
api = ["fastapi>=0.110", "uvicorn[standard]>=0.27"]
dev = ["pytest>=8", "ruff>=0.3"]

[project.scripts]
cf = "content_factory.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/content_factory"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
